refactor(#3038): isolate VoiceIdSequences into sibling module (real LoC drop)
Move the in-file `VoiceIdSequences` sub-struct (3 monotonic AtomicU64 ID
allocators) out of the `voice_barge_in.rs` giant into a new sibling module
`src/services/discord/voice_id_sequences.rs`. This is the most self-contained
remaining in-file sub-struct: 3 plain atomic counters with zero coupling to
other voice types/locks and only 3 call-sites.

- Struct + impl + its private seed constant `PROGRESS_PLAYBACK_OWNER_START`
  physically move to the sibling module.
- `INTERNAL_VOICE_MESSAGE_ID_START` stays in voice_barge_in.rs (shared with
  `is_synthetic_voice_message_id`) and is imported by the new module.
- `VoiceBargeInRuntime` keeps the `id_sequences: VoiceIdSequences` field and
  only calls the accessors; behavior, seeds, and memory Ordering preserved
  byte-for-byte. Call-sites unchanged.
- Visibility minimized to `pub(in crate::services::discord)`.

LoC: voice_barge_in.rs 4780 -> 4728 (real -52 production LoC); mod.rs 5186 ->
5187 (module registration +1). Baselines, change-surfaces freeze entry, and
generated inventory docs synced.

Gates: cargo fmt --check clean; cargo check --lib / --lib --tests clean (no
new warnings); cargo test --lib voice_barge_in (52 passed); giant ratchet
audit passed; agent-maintenance freshness passed; ci-script-checks exit 0.

Refs #3038
