+++ src/services/discord/mod.rs
@@ -93,6 +93,7 @@ mod voice_acknowledgement;
 mod voice_background_driver;
 mod voice_barge_in;
 mod voice_config_cache;
+mod voice_id_sequences;
 mod voice_routing;
 #[path = "watchers/lifecycle_decision.rs"]
 mod watcher_lifecycle_decision;
